Rename disk_api to disk
