+++ src/CMakeLists.txt
@@ -11,7 +11,7 @@ set(SOURCE_FILES
   fuse_ops.c
   keytab.c
   stego.c
-  disk_api.c
+  disk.c
 )
 
 add_compile_options(-Wall -Wextra -Wpedantic)
