+++ src/disk.h
